Refactor commonly used scope validation logic
