+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.util/src/main/java/org/wso2/carbon/apimgt/rest/api/util/impl/OAuthJwtAuthenticatorImpl.java
@@ -25,11 +25,9 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.cxf.message.Message;
 import org.wso2.carbon.apimgt.api.APIManagementException;
-import org.wso2.carbon.apimgt.api.OAuthTokenInfo;
 import org.wso2.carbon.apimgt.common.gateway.dto.JWTValidationInfo;
 import org.wso2.carbon.apimgt.common.gateway.dto.TokenIssuerDto;
 import org.wso2.carbon.apimgt.impl.APIConstants;
-import org.wso2.carbon.apimgt.impl.APIConstants.JwtTokenConstants;
 import org.wso2.carbon.apimgt.impl.RESTAPICacheConfiguration;
 import org.wso2.carbon.apimgt.impl.jwt.JWTValidator;
 import org.wso2.carbon.apimgt.impl.jwt.SignedJWTInfo;
@@ -38,13 +36,9 @@ import org.wso2.carbon.apimgt.rest.api.common.APIMConfigUtil;
 import org.wso2.carbon.apimgt.rest.api.common.RestApiConstants;
 import org.wso2.carbon.apimgt.rest.api.util.MethodStats;
 import org.wso2.carbon.apimgt.rest.api.util.authenticators.AbstractOAuthAuthenticator;
+import org.wso2.carbon.apimgt.rest.api.util.utils.OauthUtils;
 import org.wso2.carbon.apimgt.rest.api.util.utils.RestApiUtil;
-import org.wso2.carbon.context.PrivilegedCarbonContext;
-import org.wso2.carbon.user.api.UserStoreException;
-import org.wso2.carbon.user.core.service.RealmService;
-import org.wso2.carbon.utils.CarbonUtils;
 import org.wso2.carbon.utils.multitenancy.MultitenantConstants;
-import org.wso2.carbon.utils.multitenancy.MultitenantUtils;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -105,7 +99,8 @@ public class OAuthJwtAuthenticatorImpl extends AbstractOAuthAuthenticator {
                         getRESTAPITokenCache().put(jwtTokenIdentifier, jwtValidationInfo);
                     }
                     //Validating scopes
-                    return handleScopeValidation(message, signedJWTInfo, accessToken);
+                    OauthUtils oauthUtils = new OauthUtils();
+                    return oauthUtils.handleScopeValidation(message, signedJWTInfo, accessToken);
                 } else {
                     log.error("Invalid JWT token :" + maskedToken);
                     return false;
@@ -120,73 +115,6 @@ public class OAuthJwtAuthenticatorImpl extends AbstractOAuthAuthenticator {
         } catch (MalformedURLException e) {
             log.error("Malformed URL found in request path.Reason: " + e.getMessage());
         }
-        return false;
-    }
-
-    /**
-     * Handle scope validation
-     *
-     * @param accessToken   JWT token
-     * @param signedJWTInfo : Signed token info
-     * @param message       : cxf Message
-     */
-    private boolean handleScopeValidation(Message message, SignedJWTInfo signedJWTInfo, String accessToken)
-            throws APIManagementException, ParseException {
-
-        String maskedToken = message.get(RestApiConstants.MASKED_TOKEN).toString();
-        OAuthTokenInfo oauthTokenInfo = new OAuthTokenInfo();
-        oauthTokenInfo.setAccessToken(accessToken);
-        oauthTokenInfo.setEndUserName(signedJWTInfo.getJwtClaimsSet().getSubject());
-        String scopeClaim = signedJWTInfo.getJwtClaimsSet().getStringClaim(JwtTokenConstants.SCOPE);
-        if (scopeClaim != null) {
-            String orgId = RestApiUtil.resolveOrganization(message);
-            String[] scopes = scopeClaim.split(JwtTokenConstants.SCOPE_DELIMITER);
-            scopes = java.util.Arrays.stream(scopes).filter(s -> s.contains(orgId))
-                    .map(s -> s.replace(APIConstants.URN_CHOREO + orgId + ":", ""))
-                    .toArray(size -> new String[size]);
-            oauthTokenInfo.setScopes(scopes);
-
-            if (validateScopes(message, oauthTokenInfo)) {
-                //Add the user scopes list extracted from token to the cxf message
-                message.getExchange().put(RestApiConstants.USER_REST_API_SCOPES, oauthTokenInfo.getScopes());
-                //If scope validation successful then set tenant name and user name to current context
-                String tenantDomain = MultitenantUtils.getTenantDomain(oauthTokenInfo.getEndUserName());
-                int tenantId;
-                PrivilegedCarbonContext carbonContext = PrivilegedCarbonContext.getThreadLocalCarbonContext();
-                RealmService realmService = (RealmService) carbonContext.getOSGiService(RealmService.class, null);
-                try {
-                    String username = oauthTokenInfo.getEndUserName();
-                    if (MultitenantConstants.SUPER_TENANT_DOMAIN_NAME.equals(tenantDomain)) {
-                        //when the username is an email in supertenant, it has at least 2 occurrences of '@'
-                        long count = username.chars().filter(ch -> ch == '@').count();
-                        //in the case of email, there will be more than one '@'
-                        boolean isEmailUsernameEnabled = Boolean.parseBoolean(CarbonUtils.getServerConfiguration().
-                                getFirstProperty("EnableEmailUserName"));
-                        if (isEmailUsernameEnabled || (username.endsWith(SUPER_TENANT_SUFFIX) && count <= 1)) {
-                            username = MultitenantUtils.getTenantAwareUsername(username);
-                        }
-                    }
-                    if (log.isDebugEnabled()) {
-                        log.debug("username = " + username + "masked token " + maskedToken);
-                    }
-                    tenantId = realmService.getTenantManager().getTenantId(tenantDomain);
-                    carbonContext.setTenantDomain(tenantDomain);
-                    carbonContext.setTenantId(tenantId);
-                    carbonContext.setUsername(username);
-                    if (!tenantDomain.equals(MultitenantConstants.SUPER_TENANT_DOMAIN_NAME)) {
-                        APIUtil.loadTenantConfigBlockingMode(tenantDomain);
-                    }
-                    return true;
-                } catch (UserStoreException e) {
-                    log.error("Error while retrieving tenant id for tenant domain: " + tenantDomain, e);
-                }
-                log.debug("Scope validation success for the token " + maskedToken);
-                return true;
-            }
-            log.error("scopes validation failed for the token" + maskedToken);
-            return false;
-        }
-        log.error("scopes validation failed for the token" + maskedToken);
         return false;
     }
