+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.util/src/main/java/org/wso2/carbon/apimgt/rest/api/util/impl/OAuthOpaqueAuthenticatorImpl.java
@@ -28,6 +28,7 @@ import org.wso2.carbon.apimgt.impl.internal.ServiceReferenceHolder;
 import org.wso2.carbon.apimgt.impl.utils.APIUtil;
 import org.wso2.carbon.apimgt.rest.api.common.RestApiConstants;
 import org.wso2.carbon.apimgt.rest.api.util.MethodStats;
+import org.wso2.carbon.apimgt.rest.api.util.utils.OauthUtils;
 import org.wso2.carbon.apimgt.rest.api.util.utils.RestApiUtil;
 import org.wso2.carbon.context.PrivilegedCarbonContext;
 import org.wso2.carbon.identity.oauth2.OAuth2TokenValidationService;
@@ -109,7 +110,8 @@ public class OAuthOpaqueAuthenticatorImpl extends AbstractOAuthAuthenticator {
             // If token is valid then we have to do other validations and set user and tenant to carbon context.
             // Scope validation should come here.
             // If access token is valid then we will perform scope check for given resource.
-            if (validateScopes(message, tokenInfo)) {
+            OauthUtils oauthUtils = new OauthUtils();
+            if (oauthUtils.validateScopes(message, tokenInfo)) {
                 //Add the user scopes list extracted from token to the cxf message
                 message.getExchange().put(RestApiConstants.USER_REST_API_SCOPES, tokenInfo.getScopes());
                 //If scope validation successful then set tenant name and user name to current context
