+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.util/src/main/java/org/wso2/carbon/apimgt/rest/api/util/utils/OauthUtils.java
@@ -1,0 +1,198 @@
+/*
+ *  Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.apimgt.rest.api.util.utils;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.apache.cxf.message.Message;
+import org.wso2.carbon.apimgt.api.APIManagementException;
+import org.wso2.carbon.apimgt.api.OAuthTokenInfo;
+import org.wso2.carbon.apimgt.api.model.Scope;
+import org.wso2.carbon.apimgt.api.model.URITemplate;
+import org.wso2.carbon.apimgt.impl.APIConstants;
+import org.wso2.carbon.apimgt.impl.jwt.SignedJWTInfo;
+import org.wso2.carbon.apimgt.impl.utils.APIUtil;
+import org.wso2.carbon.apimgt.rest.api.common.RestApiConstants;
+import org.wso2.carbon.apimgt.rest.api.util.MethodStats;
+import org.wso2.carbon.context.PrivilegedCarbonContext;
+import org.wso2.carbon.user.api.UserStoreException;
+import org.wso2.carbon.user.core.service.RealmService;
+import org.wso2.carbon.utils.CarbonUtils;
+import org.wso2.carbon.utils.multitenancy.MultitenantConstants;
+import org.wso2.carbon.utils.multitenancy.MultitenantUtils;
+import org.wso2.uri.template.URITemplateException;
+
+import java.text.ParseException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+public class OauthUtils {
+
+    private static final Log log = LogFactory.getLog(OauthUtils.class);
+    private static final String SUPER_TENANT_SUFFIX =
+            APIConstants.EMAIL_DOMAIN_SEPARATOR + MultitenantConstants.SUPER_TENANT_DOMAIN_NAME;
+
+    /**
+     * Handle scope validation
+     *
+     * @param accessToken   JWT token
+     * @param signedJWTInfo : Signed token info
+     * @param message       : cxf Message
+     */
+    public boolean handleScopeValidation(Message message, SignedJWTInfo signedJWTInfo, String accessToken)
+            throws APIManagementException, ParseException {
+
+        String maskedToken = message.get(RestApiConstants.MASKED_TOKEN).toString();
+        OAuthTokenInfo oauthTokenInfo = new OAuthTokenInfo();
+        oauthTokenInfo.setAccessToken(accessToken);
+        oauthTokenInfo.setEndUserName(signedJWTInfo.getJwtClaimsSet().getSubject());
+        String scopeClaim = signedJWTInfo.getJwtClaimsSet().getStringClaim(APIConstants.JwtTokenConstants.SCOPE);
+        if (scopeClaim != null) {
+            String orgId = RestApiUtil.resolveOrganization(message);
+            String[] scopes = scopeClaim.split(APIConstants.JwtTokenConstants.SCOPE_DELIMITER);
+            scopes = java.util.Arrays.stream(scopes).filter(s -> s.contains(orgId))
+                    .map(s -> s.replace(APIConstants.URN_CHOREO + orgId + ":", ""))
+                    .toArray(size -> new String[size]);
+            oauthTokenInfo.setScopes(scopes);
+            if (validateScopes(message, oauthTokenInfo)) {
+                //Add the user scopes list extracted from token to the cxf message
+                message.getExchange().put(RestApiConstants.USER_REST_API_SCOPES, oauthTokenInfo.getScopes());
+                //If scope validation successful then set tenant name and user name to current context
+                String tenantDomain = MultitenantUtils.getTenantDomain(oauthTokenInfo.getEndUserName());
+                int tenantId;
+                PrivilegedCarbonContext carbonContext = PrivilegedCarbonContext.getThreadLocalCarbonContext();
+                RealmService realmService = (RealmService) carbonContext.getOSGiService(RealmService.class, null);
+                try {
+                    String username = oauthTokenInfo.getEndUserName();
+                    if (MultitenantConstants.SUPER_TENANT_DOMAIN_NAME.equals(tenantDomain)) {
+                        //when the username is an email in supertenant, it has at least 2 occurrences of '@'
+                        long count = username.chars().filter(ch -> ch == '@').count();
+                        //in the case of email, there will be more than one '@'
+                        boolean isEmailUsernameEnabled = Boolean.parseBoolean(CarbonUtils.getServerConfiguration().
+                                getFirstProperty("EnableEmailUserName"));
+                        if (isEmailUsernameEnabled || (username.endsWith(SUPER_TENANT_SUFFIX) && count <= 1)) {
+                            username = MultitenantUtils.getTenantAwareUsername(username);
+                        }
+                    }
+                    if (log.isDebugEnabled()) {
+                        log.debug("username = " + username + "masked token " + maskedToken);
+                    }
+                    tenantId = realmService.getTenantManager().getTenantId(tenantDomain);
+                    carbonContext.setTenantDomain(tenantDomain);
+                    carbonContext.setTenantId(tenantId);
+                    carbonContext.setUsername(username);
+                    if (!tenantDomain.equals(MultitenantConstants.SUPER_TENANT_DOMAIN_NAME)) {
+                        APIUtil.loadTenantConfigBlockingMode(tenantDomain);
+                    }
+                    return true;
+                } catch (UserStoreException e) {
+                    log.error("Error while retrieving tenant id for tenant domain: " + tenantDomain, e);
+                }
+                log.debug("Scope validation success for the token " + maskedToken);
+                return true;
+            }
+            log.error("scopes validation failed for the token" + maskedToken);
+            return false;
+        }
+        log.error("scopes validation failed for the token" + maskedToken);
+        return false;
+    }
+
+    /**
+     * @param message   CXF message to be validate
+     * @param tokenInfo Token information associated with incoming request
+     * @return return true if we found matching scope in resource and token information
+     * else false(means scope validation failed).
+     */
+    @MethodStats
+    public boolean validateScopes(Message message, OAuthTokenInfo tokenInfo) {
+        String basePath = (String) message.get(Message.BASE_PATH);
+        // path is obtained from Message.REQUEST_URI instead of Message.PATH_INFO, as Message.PATH_INFO contains
+        // decoded values of request parameters
+        String path = (String) message.get(Message.REQUEST_URI);
+        String verb = (String) message.get(Message.HTTP_REQUEST_METHOD);
+        String resource = path.substring(basePath.length() - 1);
+        String[] scopes = tokenInfo.getScopes();
+
+        String version = (String) message.get(RestApiConstants.API_VERSION);
+
+        //get all the URI templates of the REST API from the base path
+        Set<URITemplate> uriTemplates = RestApiUtil.getURITemplatesForBasePath(basePath + version);
+        if (uriTemplates.isEmpty()) {
+            if (log.isDebugEnabled()) {
+                log.debug("No matching scopes found for request with path: " + basePath
+                        + ". Skipping scope validation.");
+            }
+            return true;
+        }
+
+        for (Object template : uriTemplates.toArray()) {
+            org.wso2.uri.template.URITemplate templateToValidate = null;
+            Map<String, String> var = new HashMap<String, String>();
+            //check scopes with what we have
+            String templateString = ((URITemplate) template).getUriTemplate();
+            try {
+                templateToValidate = new org.wso2.uri.template.URITemplate(templateString);
+            } catch (URITemplateException e) {
+                log.error("Error while creating URI Template object to validate request. Template pattern: " +
+                        templateString, e);
+            }
+            if (templateToValidate != null && templateToValidate.matches(resource, var) && scopes != null
+                    && verb != null && verb.equalsIgnoreCase(((URITemplate) template).getHTTPVerb())) {
+                for (String scope : scopes) {
+                    Scope scp = ((URITemplate) template).getScope();
+                    if (scp != null) {
+                        if (scope.equalsIgnoreCase(scp.getKey())) {
+                            //we found scopes matches
+                            if (log.isDebugEnabled()) {
+                                log.debug("Scope validation successful for access token: " +
+                                        message.get(RestApiConstants.MASKED_TOKEN) + " with scope: " + scp.getKey() +
+                                        " for resource path: " + path + " and verb " + verb);
+                            }
+                            return true;
+                        }
+                    } else if (!((URITemplate) template).retrieveAllScopes().isEmpty()) {
+                        List<Scope> scopesList = ((URITemplate) template).retrieveAllScopes();
+                        for (Scope scpObj : scopesList) {
+                            if (scope.equalsIgnoreCase(scpObj.getKey())) {
+                                //we found scopes matches
+                                if (log.isDebugEnabled()) {
+                                    log.debug("Scope validation successful for access token: " +
+                                            message.get(RestApiConstants.MASKED_TOKEN) + " with scope: " + scpObj.getKey() +
+                                            " for resource path: " + path + " and verb " + verb);
+                                }
+                                return true;
+                            }
+                        }
+                    } else {
+                        if (log.isDebugEnabled()) {
+                            log.debug("Scope not defined in swagger for matching resource " + resource + " and verb "
+                                    + verb + " . So consider as anonymous permission and let request to continue.");
+                        }
+                        return true;
+                    }
+                }
+            }
+        }
+        return false;
+    }
+
+}
